@@ -1,11 +1,11 @@
 {
     "name": "kukrik/toastr",
-    "description": "QCubed plugin for Toastr",
+    "description": "QCubed-4 plugin for Toastr",
     "type": "qcubed-library",
     "keywords": ["qcubed", "toastr"],
     "license": "MIT",
     "require": {
-        "php": ">=5.6"
+        "php": ">=8.3"
     },
     "autoload": {
       "psr-4": {"QCubed\\Plugin\\": "src/"}
